% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backtraking.R
\name{backtraking}
\alias{backtraking}
\title{Algorithme de Backtraking}
\usage{
backtraking(matrice_vide)
}
\arguments{
\item{matrice_vide}{: sudoku à compléter}
}
\value{
le competur
}
\description{
Algorithme de Backtraking
}
