% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bloc.R
\name{bloc}
\alias{bloc}
\title{Fonction qui retourne le numéro du bloc d'une case donnée d'une matrice 9x9
la matrice du sudoku est de la forme:
 1 2 3
 4 5 6
 7 8 9}
\usage{
bloc(i, j)
}
\arguments{
\item{i}{: la ligne de la case dans la matrice}

\item{j}{: la colonne de la case dans la matrice}
}
\value{
le numéro du bloc d'une case
}
\description{
Fonction qui retourne le numéro du bloc d'une case donnée d'une matrice 9x9
la matrice du sudoku est de la forme:
 1 2 3
 4 5 6
 7 8 9
}
