% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_possibilite.R
\name{count_pos}
\alias{count_pos}
\title{Fonction qui retourne une matrice avec le nombre de valeurs possibles à mettre par case vide,
et si la case est non-vide alors dans la matrice retournée, cette case est vide.}
\usage{
count_pos(matrice)
}
\arguments{
\item{matrice}{: matrice 9x9 avec des cases vides, c'est la grille de sudoku à compléter}
}
\value{
une matrice complétée avec le nombre de possibilité par case
}
\description{
Fonction qui retourne une matrice avec le nombre de valeurs possibles à mettre par case vide,
et si la case est non-vide alors dans la matrice retournée, cette case est vide.
}
