% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_complet.R
\name{est_complet}
\alias{est_complet}
\title{Fonction qui retourne TRUE si la matrice est complète, ie le sudoku est complet}
\usage{
est_complet(matrice)
}
\arguments{
\item{matrice}{: matrice 9x9}
}
\value{
TRUE or FALSE
}
\description{
Fonction qui retourne TRUE si la matrice est complète, ie le sudoku est complet
}
