% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estcolonne.R
\name{estcolonne}
\alias{estcolonne}
\title{Fonction qui retourne TRUE si une valeur donnée est dans une colonne donnée, FALSE sinon}
\usage{
estcolonne(val, j, matrice)
}
\arguments{
\item{val}{: la valeur à tester}

\item{j}{: la colonne où l'on veut regarder si il y a la valeur}

\item{matrice_vide}{: matrice 9x9 avec des cases vides, c'est la grille de sudoku à compléter}
}
\value{
TRUE or FALSE
}
\description{
Fonction qui retourne TRUE si une valeur donnée est dans une colonne donnée, FALSE sinon
}
