% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estdansbloc.R
\name{estdansbloc}
\alias{estdansbloc}
\title{Fonction qui retourne TRUE si une valeur donnée et dans un bloc choisi, FALSE sinon}
\usage{
estdansbloc(val, b, matrice)
}
\arguments{
\item{val}{: la valeur à tester}

\item{b}{: le bloc où l'on veut regarder si il y a la valeur}

\item{matrice_vide}{: matrice 9x9 avec des cases vides, c'est la grille de sudoku à compléter}
}
\value{
TRUE or FALSE
}
\description{
Fonction qui retourne TRUE si une valeur donnée et dans un bloc choisi, FALSE sinon
}
