% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estligne.R
\name{estligne}
\alias{estligne}
\title{Fonction qui retourne TRUE si une valeur donnée est dans une ligne donnée, FALSE sinon}
\usage{
estligne(val, i, matrice)
}
\arguments{
\item{val}{: la valeur à tester}

\item{i}{: la ligne où l'on veut regarder si il y a la valeur}

\item{matrice_vide}{: matrice 9x9 avec des cases vides, c'est la grille de sudoku à compléter}
}
\value{
TRUE or FALSE
}
\description{
Fonction qui retourne TRUE si une valeur donnée est dans une ligne donnée, FALSE sinon
}
