% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Affiche_grille.R
\name{grille}
\alias{grille}
\title{Affiche un grille de sudoku}
\usage{
grille(MAT)
}
\arguments{
\item{matrice}{: matrice de taille 9x9}
}
\value{
affiche une matrice avec la fonction 'grid.table' du package gridExtra
}
\description{
Affiche un grille de sudoku
}
