% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indice_vect.R
\name{indice_vect}
\alias{indice_vect}
\title{Fonction qui retourne l'indice d'une valeur donnée dans un vecteur donné}
\usage{
indice_vect(valeur, vecteur_possib)
}
\arguments{
\item{valeur}{: la valeur à tester}

\item{vecteur_possib}{: vecteur contenant que des chiffres}
}
\value{
l'indice de la valeur dans le vecteur, si la valeur
n'appartient pas au vecteur alors la fonction retourne indice = 0
}
\description{
Fonction qui retourne l'indice d'une valeur donnée dans un vecteur donné
}
