% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nb_case_vide.R
\name{nb_case_vide}
\alias{nb_case_vide}
\title{Fonction qui retourne le nombre de cases vides dans une matrice donnée, ie une grille de sudoku}
\usage{
nb_case_vide(matrice_vide)
}
\arguments{
\item{matrice_vide}{: matrice 9x9, c'est le sudoku à compléter}
}
\value{
le nombre de cases vides dans la matrice
}
\description{
Fonction qui retourne le nombre de cases vides dans une matrice donnée, ie une grille de sudoku
}
