% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nb_possib_case.R
\name{nbpossib_case}
\alias{nbpossib_case}
\title{Fonction qui retourne le nombre de valeurs possibles à mettre dans une case}
\usage{
nbpossib_case(i, j, matrice_vide)
}
\arguments{
\item{i}{: ligne de la case}

\item{j}{: colonne de la case}

\item{matrice_vide}{: matrice 9x9, c'est la grille de sudoku à compléter}
}
\value{
le nombre de retour effectué pour résoudre le sudoku
}
\description{
Fonction qui retourne le nombre de valeurs possibles à mettre dans une case
}
