% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permute.R
\name{permute}
\alias{permute}
\title{Fonction qui permute aléatoirement les lignes et les colonnes d'un même bloc}
\usage{
permute(matrice_complet)
}
\arguments{
\item{matrice_complet}{: matrice 9x9 complète}
}
\value{
Affiche la nouvelle matrice avec ses lignes et ses colonnes permutées aléatoirement
}
\description{
Fonction qui permute aléatoirement les lignes et les colonnes d'un même bloc
}
