% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sudoku_grille_vide.R
\name{SUDOKU}
\alias{SUDOKU}
\title{Fonction qui ne prend aucun paramètre en entrée, et qui affiche une grille de
sudoku à compléter, c'est à dire une matrice, avec la fonction 'grid.table'}
\usage{
SUDOKU()
}
\description{
Fonction qui ne prend aucun paramètre en entrée, et qui affiche une grille de
sudoku à compléter, c'est à dire une matrice, avec la fonction 'grid.table'
}
