% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tableau.R
\name{tableau}
\alias{tableau}
\title{Fonction qui retourne un data-frame de taille (n x p) :
avec n = le nombre de cases vides dans la matrice d'entrée et p = 3
Dans chaque ligne du data-frame, on retrouve les coordonnées (ligne et colonne) d'une case,
puis dans la 3ème colonne, on a le nombre de possibilité de cette case.
Les lignes de ce data-frame sont ordonnées par le nombre de possibilité de la case, de plus
petit au plus grand. C'est à dire, ce data-frame est ordonné par ordre croissant sur les
valeurs de sa 3ème colonne}
\usage{
tableau(matrice_vide)
}
\arguments{
\item{matrice_vide}{: matrice 9x9 avec des cases vides, c'est la grille de sudoku à compléter}
}
\value{
un data frame de taille (n x p) :
avec n = le nombre de cases vides dans la matrice d'entrée et p = 3
}
\description{
Fonction qui retourne un data-frame de taille (n x p) :
avec n = le nombre de cases vides dans la matrice d'entrée et p = 3
Dans chaque ligne du data-frame, on retrouve les coordonnées (ligne et colonne) d'une case,
puis dans la 3ème colonne, on a le nombre de possibilité de cette case.
Les lignes de ce data-frame sont ordonnées par le nombre de possibilité de la case, de plus
petit au plus grand. C'est à dire, ce data-frame est ordonné par ordre croissant sur les
valeurs de sa 3ème colonne
}
