% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tri_possibilite.R
\name{tri_possibilite}
\alias{tri_possibilite}
\title{Fonction qui prend en entrée une matrice avec, dans les cases vides du sudoku à compléter,
le nombre de possibilités par case. Et qui retourne une matrice qui ordonne par ordre
croissant ces cases.}
\usage{
tri_possibilite(count_matrice)
}
\arguments{
\item{count_matrice}{: matrice 9x9 avec dans les cases qui sont à compléter}
}
\value{
Affiche la matrice avec les ordres des cases triés.
}
\description{
Fonction qui prend en entrée une matrice avec, dans les cases vides du sudoku à compléter,
le nombre de possibilités par case. Et qui retourne une matrice qui ordonne par ordre
croissant ces cases.
}
