% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Val_possible.R
\name{val_possible}
\alias{val_possible}
\title{Fonction qui retourne TRUE, si une valeur est possible dans une case, et FALSE sinon}
\usage{
val_possible(val, i, j, matrice_vide)
}
\arguments{
\item{val}{: valeur à tester}

\item{i}{: ligne de la case}

\item{j}{: colonne de la case}

\item{matrice_vide}{: matrice 9x9 avec des cases vides, c'est la grille de sudoku à compléter}
}
\value{
TRUE si la on peut mettre la valeur dans la case donnée, FALSE sinon
}
\description{
Fonction qui retourne TRUE, si une valeur est possible dans une case, et FALSE sinon
}
