% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vide.R
\name{vide}
\alias{vide}
\title{Fonction qui retourne une matrice avec des cases vidés aléatoirement}
\usage{
vide(matrice_complet)
}
\arguments{
\item{matrice_complet}{: matrice 9x9 complète}
}
\value{
la matrice 9x9 avec des cases vides
}
\description{
Fonction qui retourne une matrice avec des cases vidés aléatoirement
}
