% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backtraking.R
\name{backtraking}
\alias{backtraking}
\title{Algorithme de Backtraking}
\usage{
backtraking(matrice_vide)
}
\arguments{
\item{matrice_vide}{: matrice 9x9 avec des cases vides, c'est la grille de sudoku à compléter}
}
\value{
compteur = le nombre de retour effectué pour résoudre le sudoku
}
\description{
Algorithme de Backtraking
}
