% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Possibilite_vecteur.R
\name{possibilite}
\alias{possibilite}
\title{Fonction}
\usage{
possibilite(i, j, matrice_vide)
}
\arguments{
\item{matrice_vide}{: sudoku à compléter}
}
\value{
le nombre de retour effectué pour résoudre le sudoku
}
\description{
Fonction
}
